requirements.txt
# GUI Framework
PySide6>=6.5.0

# PDF Tools
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
pikepdf>=8.0.0
reportlab>=4.0.0

# QR & Barcode Tools
qrcode>=7.4.2
pyzbar>=0.1.9
python-barcode>=0.15.1

# Image Tools
Pillow>=10.0.0
piexif>=1.1.3

# File Conversion Tools
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
docx2pdf>=0.1.8

# System Tools
cryptography>=41.0.0

# Network Tools
requests>=2.31.0
speedtest-cli>=2.1.3

# API Framework (for Railway deployment)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Build Tools
pyinstaller>=6.0.0

# Additional Tools
click>=8.1.0
packaging>=23.0